# openapi.yaml
openapi: 3.0.0
info:
  title: AFOCE Accounting API
  description: |
    Complete accounting API for invoicing, expenses, and financial management.
    Supports multi-tenant workspaces with role-based access control.
  version: 1.0.0
  contact:
    name: AFOCE Support
    email: support@afoce.app
servers:
  - url: https://api.afoce.app
    description: Production server
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
paths:
...
  /api/expenses:
    get:
      summary: List expenses
      operationId: listExpenses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_approval, manager_review, blocked, approved, rejected]
      responses:
        "200":
          description: List of expenses
    post:
      summary: Create expense
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Expense created
  /api/expenses/{id}/approve:
    post:
      summary: Approve/reject expense
      operationId: approveExpense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [approve, reject, request_review]
                notes:
                  type: string
      responses:
        "200":
          description: Expense processed
  /api/batch:
    post:
      summary: Batch operations
      operationId: batchOperation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRequest"
      responses:
        "200":
          description: Batch operation completed
    patch:
      summary: Batch status update
      operationId: batchStatusUpdate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchStatusRequest"
      responses:
        "200":
          description: Status updated
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: supabase-auth-token
  schemas:
...
    CreateExpenseRequest:
      type: object
      required:
        - employee
        - category
        - amount
        - bs_date
        - ad_date
      properties:
        employee:
          type: string
        category:
          type: string
        amount:
          type: number
        bs_date:
          type: string
        ad_date:
          type: string
        receipt_url:
          type: string
...
    BatchRequest:
      type: object
      properties:
        operation:
          type: string
          enum: [create, update, delete]
        entity:
          type: string
        items:
          type: array
    BatchStatusRequest:
      type: object
      properties:
        entity:
          type: string
        ids:
          type: array
          items:
            type: string
        status:
          type: string
        reason:
          type: string